301_prefect/sample2/requirements.txt
# requirements.txt

# --- Core Libraries (Versions pinned for reproducibility) ---
prefect==2.19.4
duckdb==0.10.3
PyYAML==6.0.1
Jinja2==3.1.4
pandas==2.2.2

# --- Dependency Fixes ---
# Prefect/Pydantic compatibility issues.
pydantic==2.6.4

# Prefect/griffe compatibility issues.
griffe[toml]==0.38.1